requirements.txt
# YouTube Playlist to Markdown - Python Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
python-dotenv>=1.0.0
google-generativeai>=0.8.0
youtube-transcript-api>=0.1.0

# Optional but recommended for better compatibility
requests>=2.28.0
urllib3>=1.26.0,<3.0.0